Release 0.2.9

+++ setup.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 #
 # #             #
 # #mmm   mmmm   #   m   mmmm
@@ -9,14 +9,17 @@
 #        "               ""
 #        bash package manager
 
+VERSION="0.2.9"
+TAG=${TAG:-$VERSION}
+BRANCH=${BRANCH:-$TAG}
 REMOTE=${REMOTE:-https://github.com/bpkg/bpkg.git}
 TMPDIR=${TMPDIR:-/tmp}
-DEST=${DEST:-${TMPDIR}/bpkg-master}
+DEST=${DEST:-$TMPDIR/bpkg-$BRANCH}
 
 ## test if command exists
 ftest () {
-  echo "  info: Checking for ${1}..."
-  if ! type -f "${1}" > /dev/null 2>&1; then
+  echo "  info: Checking for $1..."
+  if ! type -f "$1" > /dev/null 2>&1; then
     return 1
   else
     return 0
@@ -26,8 +29,8 @@ ftest () {
 ## feature tests
 features () {
   for f in "${@}"; do
-    ftest "${f}" || {
-      echo >&2 "  error: Missing \`${f}'! Make sure it exists and try again."
+    ftest "$f" || {
+      echo >&2 "  error: Missing \`$f'! Make sure it exists and try again."
       return 1
     }
   done
@@ -43,13 +46,15 @@ setup () {
   ## build
   {
     echo
-    cd "${TMPDIR}"
     echo "  info: Creating temporary files..."
-    test -d "${DEST}" && { echo "  warn: Already exists: '${DEST}'"; }
-    rm -rf "${DEST}"
-    echo "  info: Fetching latest 'bpkg'..."
-    git clone --depth=1 "${REMOTE}" "${DEST}" > /dev/null 2>&1
-    cd "${DEST}"
+    cd "$TMPDIR" || exit
+    test -d "$DEST" && { echo "  warn: Already exists: '$DEST'"; }
+    rm -rf "$DEST"
+
+    echo "  info: Fetching 'bpkg@$BRANCH'..."
+    git clone --depth=1 --branch "$BRANCH" "$REMOTE" "$DEST" > /dev/null 2>&1
+    cd "$DEST" || exit
+
     echo "  info: Installing..."
     echo
     make_install
@@ -60,39 +65,86 @@ setup () {
 
 ## make targets
 BIN="bpkg"
-[ -z "$PREFIX" ] && PREFIX="/usr/local"
+if [ -z "$PREFIX" ]; then
+  if [ "$(whoami)" == "root" ]; then
+    PREFIX="/usr/local"
+  else
+    PREFIX="$HOME/.local"
+  fi
+fi
 
 # All 'bpkg' supported commands
-CMDS="json install package term suggest init utils update list show getdeps"
+declare -a CMDS=()
+CMDS+=("env")
+CMDS+=("getdeps")
+CMDS+=("init")
+CMDS+=("install")
+CMDS+=("json")
+CMDS+=("list")
+CMDS+=("package")
+CMDS+=("run")
+CMDS+=("show")
+CMDS+=("source")
+CMDS+=("suggest")
+CMDS+=("term")
+CMDS+=("update")
+CMDS+=("utils")
+CMDS+=("realpath")
 
 make_install () {
+  local source
+
+  ## do 'make uninstall'
   make_uninstall
+
   echo "  info: Installing $PREFIX/bin/$BIN..."
   install -d "$PREFIX/bin"
-  local source=$(<$BIN)
-  [ -f "$source" ] && install "$source" "$PREFIX/bin/$BIN" || install "$BIN" "$PREFIX/bin"
-  for cmd in $CMDS; do
-    source=$(<$BIN-$cmd)
-    [ -f "$source" ] && install "$source" "$PREFIX/bin/$BIN-$cmd" || install "$BIN-$cmd" "$PREFIX/bin"
+  source=$(<$BIN)
+
+  if [ -f "$source" ]; then
+    install "$source" "$PREFIX/bin/$BIN"
+  else
+    install "$BIN" "$PREFIX/bin"
+  fi
+
+  for cmd in "${CMDS[@]}"; do
+    if test -f "$BIN-$cmd"; then
+      source=$(<"$BIN-$cmd")
+
+      if [ -f "$source" ]; then
+        install "$source" "$PREFIX/bin/$BIN-$cmd"
+      else
+        install "$BIN-$cmd" "$PREFIX/bin"
+      fi
+    fi
+
   done
   return $?
 }
 
 make_uninstall () {
-  echo "  info: Uninstalling $PREFIX/bin/$BIN..."
+  echo "  info: Uninstalling $PREFIX/bin/$BIN*"
+  echo "    rm: $PREFIX/bin/$BIN'"
   rm -f "$PREFIX/bin/$BIN"
-  for cmd in $CMDS; do
-    rm -f "$PREFIX/bin/$BIN-$cmd"
+  for cmd in "${CMDS[@]}"; do
+    if test -f "$PREFIX/bin/$BIN-$cmd"; then
+      echo "    rm: $PREFIX/bin/$BIN-$cmd'"
+      rm -f "$PREFIX/bin/$BIN-$cmd"
+    fi
   done
   return $?
 }
 
 make_link () {
   make_uninstall
-  echo "  info: Linking $PREFIX/bin/$BIN..."
+  echo "  info: Linking $PREFIX/bin/$BIN*"
+  echo "  link: '$PWD/$BIN' -> '$PREFIX/bin/$BIN'"
   ln -s "$PWD/$BIN" "$PREFIX/bin/$BIN"
-  for cmd in $CMDS; do
-    ln -s "$PWD/$BIN-$cmd" "$PREFIX/bin"
+  for cmd in "${CMDS[@]}"; do
+    if test -f "$PWD/$BIN-$cmd"; then
+      echo "  link: '$PWD/$BIN-$cmd' -> '$PREFIX/bin/$BIN-$cmd'"
+      ln -s "$PWD/$BIN-$cmd" "$PREFIX/bin/$BIN-$cmd"
+    fi
   done
   return $?
 }
@@ -101,6 +153,11 @@ make_unlink () {
   make_uninstall
 }
 
-## go
-[ $# -eq 0 ] && setup || make_$1
+## do setup or `make_{install|uninstall|link|unlink}` command
+if [ $# -eq 0 ]; then
+  setup
+else
+  "make_$1"
+fi
+
 exit $?
